dmon: Implement timeout for process running time
When a timeout is specified, with "-t SECONDS", then if the command takes
longer to complete, it will be terminated by sending the SIGTERM/SIGCONT
combo, and then respawned. This can be used to run processes that must be
enforced to keep their running time below a given value.

As a bonus, suffixes for the argument to "-t" may be given: (w)eeks, (d)ays,
(h)ours and (m)inutes.