dmon: Do not wait the first time a process is started
Now dmon keeps track of the moment at which a command was started to
determine whether to sleep for 1s before respawning. The initial value
of the variable makes dmon to *not* sleep before executing commands for
first time. This is intended behavior. Using some kind of back-off for
respawns would be nice, but that would involve having an additional
counter, so for the moment this solution is good enough.