Support reading options from files with "-C file"
This implements very basic support for reading options from files.
Note that this reads "options from files", it does not read "proper
configuration files", so implementation is somewhat dodgy -- but it
the most we can do with the less possible code. It works this way:

- If "-C" is exactly the first argument and there are at least two
	arguments, the second one is expected to be a file name.
- Arguments are shifted two positions to the left.
- Contents from the file are placed in the argv array by using
	replace_args_file().
- Contents of the file will be happily "parsed" by getopt().

This is similar to what is done with the contents of the DMON_OPTIONS
environment variable.