Build a multicall binary by default
In general there is no gain in building separate binaries for each tool, so
it sounds reasonable to assume that by default everybody would like to build
a multicall binary. Still, it is possible to build standalone tools passing
MULTICALL=0 in the Make command line. The README is updated accordingly.