dlog: Add --prefix, to add a prefix text to messages
This implements inserting an arbitrary string in between timestamps and
logged message lines. This allows for some neat tricks like sending several
outputs to a FIFO, and having a different prefix for them. Example:

  mkfifo logpipe
  dmon cat logpipe -- drlog -t /var/log/combined
  dmon command1 -- dlog -p command1: logpipe
  dmon command2 -- dlog -p command2: logpipe
  dmon command3 -- dlog -p command3: logpipe

This may be interesting in some particular usage scenarios, or for
aggregating logs of several services which seldom log messages.