Re-mount the root file system read-write
This is neded as first step of dinit_init_filesystem(), so directories for
mount points can be created as needed. This applies to root file systems
mounted as read-only by the kernel (either because that is the default, or
because the "ro" option has been passed in the command line).