Implement flag to change the work directory
This adds a new command line -W/--work-dir option which allows changing
the current working directory right after reading the configuration
file. This means that the path of the configuration file is interpreted
relative to the working directory when dmon is invoked, and the rest of
paths (e.g. path to the PID file) are relative to the working directory
configured with -W/--work-dir.