First baby steps into a potential port to Rust
The followed approach is to compile a static library from the Rust code,
link it with the "dmon" binary, and move over functions one by one from
C to Rust. For more details and inspiration, check:

  https://github.com/carols10cents/rust-out-your-c-talk
  https://people.gnome.org/~federico/news-2016-10.html#25

This commit provides the basic infrastructure, and a couple of simple
functions moved from task.c to src/dmon.rs.