CI: Pass -fcommon to the compiler to avoid build failures
Pass -fcommon to the compiler to fix builds. This avoids compilation
failures due to multiple definitions of common symbols in libwheel. GCC
has recently switched to having -fno-common set by default, which is why
builds started failing. Once libwheel is not used anymore it should be
possible to revert this change.