Makefile: Bring back setting MULTICALL= in the command line
This allows passing again the MULTICALL=0|1 value in the Make command
line, without needing to use the "make standalone" or "make multicall"
targets.

In order to achieve the above, phony targets for both build modes are
introduced, and one of the other chosen depending on the value of the
$(MULTICALL) variable, which must strictly have the values "0" or "1"
(which seems to be a reasonable restriction). For example:

  target-0:
      @echo 'Handle target for a standalone build'

  target-1:
      @echo 'Handle target for a multicall build'

  target: target-$(MULTICALL)

  .PHONY: target target-0 target-1

Another trick is that variables can be used in target names, and even if
the expansion results in an empty string, it is not a syntax error, but
instead the rule is discarded:

  empty =

  $(empty): dependency  # Not a syntax error, rule is discarded.
  target: $(empty)      # Ditto.
  $(empty): $(empty)    # Ditto.

The above behavior is (ab)used to generate rules for either building
binaries or making symlinks, in conjunction with replacement expansions.
In the next example, if $A is undefined (or empty) it will be discarded,
otherwise it expands to a multi-target rule which creates symlinks:

  $A: dmon
      ln -sf dmon $@

While the following will create a multi-target rule for building each
one of the standalone programs, but only if $P is non-empty (as usual,
otherwise the rule is ignored):

  $P: $(P:=.o) libdmon.a
      $(CC) ...

Which in turn is equivalent to the following, for a value of $P equal
to "a b":

  a: a.o libdmon.a
      $(CC) ...

  b: b.o libdmon.a
      $(CC) ...

The solution works both for the BSD and GNU Make implementations. While
it should work with any implementation compatible with AT&T Unix Make,
but others have not been tested.