Prevent logging from exiting on empty lines
Before this commit, `freadline` did not distinguish between an empty
line and EOF. This would case a printf such as

```
printf("\nHello, World\n");
```

to cause dlog, dslog, or drlog to exit.

In order to fix this, this commit includes the delimiter in the result
of `freaduntil`, which means `\n` will be included in all calls to
`freadline`.

Fixes #10
Fixes #12